+++ common/core/src/main/java/module-info.java
@@ -20,6 +20,7 @@
  * Common classes used throughout, such as:
  * <li>{@link is.codion.common.observer.Observer}
  * <li>{@link is.codion.common.observer.Observable}
+ * <li>{@link is.codion.common.observer.Mutable}
  * <li>{@link is.codion.common.event.Event}
  * <li>{@link is.codion.common.state.State}
  * <li>{@link is.codion.common.state.StateObserver}